requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
openpyxl>=3.0.0
PyYAML>=6.0
jupyter>=1.0.0
notebook>=6.4.0